Create stable-diffusion-webui-directml
"Adding stable-diffusion-webui-directml to improve GPU utilization"

You can commit this change to the stable-diffusion-webui-docker repository. This will record the fact that you've added the stable-diffusion-webui-directml submodule and which commit it's currently at. Other people who clone your stable-diffusion-webui-docker repository will be able to initialize and update the stable-diffusion-webui-directml submodule to get the same code you're using.